Move text into screen directory
