+++ include/radix/screen/XmlScreenLoader.hpp
@@ -7,7 +7,7 @@
 
 #include <radix/screen/Screen.hpp>
 #include <radix/util/XmlLoader.hpp>
-#include <radix/text/Text.hpp>
+#include <radix/screen/text/Text.hpp>
 
 namespace radix {
